Run with multiple instances

+++ src/main/kotlin/no/nav/nks_ai/core/ConversationDeletionJob.kt
@@ -3,20 +3,20 @@ package no.nav.nks_ai.core
 import dev.starry.ktscheduler.job.Job
 import dev.starry.ktscheduler.scheduler.KtScheduler
 import dev.starry.ktscheduler.triggers.DailyTrigger
+import io.github.oshai.kotlinlogging.KotlinLogging
+import io.ktor.client.HttpClient
+import no.nav.nks_ai.app.isLeader
 import no.nav.nks_ai.core.conversation.ConversationService
 
+private val logger = KotlinLogging.logger { }
+
 class ConversationDeletionJob(
-    private val conversationService: ConversationService
+    private val conversationService: ConversationService,
+    private val httpClient: HttpClient,
 ) {
     fun start() {
-        if (isLeader()) {
-            scheduleJob()
-        }
+        scheduleJob()
     }
-
-    // TODO must be implemented before running multiple pods.
-    //      See: https://doc.nais.io/services/leader-election/how-to/enable/#__tabbed_1_1
-    private fun isLeader() = true
 
     // Deletes old conversations at 03:00 every night.
     private fun scheduleJob() {
@@ -26,7 +26,13 @@ class ConversationDeletionJob(
             jobId = "DeleteOldConversations",
             trigger = trigger,
             callback = {
-                conversationService.deleteOldConversations()
+                if (isLeader(httpClient)) {
+                    conversationService.deleteOldConversations()
+                } else {
+                    logger.info {
+                        "This instance is not leader. Scheduled job will not be performed"
+                    }
+                }
             }
         )
